pyproject.toml
[project]
name = "reolink-ctl"
version = "0.2.0"
description = "CLI for Reolink camera control"
requires-python = ">=3.9"
dependencies = [
    "reolink_aio>=0.9.0,<0.9.8",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
reolink-ctl = "reolink_ctl.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
